+++ kernel/include/kernel/parallel/parallel_algos/linear_algebra/dot_product.h
@@ -7,10 +7,10 @@
   */
 
 #include "kernel/base/types.h"
-#include "kernel/executors/simple_task.h"
-#include "kernel/executors/task_uitilities.h"
-#include "kernel/models/reduction_operations.h"
-#include "kernel/utilities/result_holder.h"
+#include "kernel/parallel/threading/simple_task.h"
+#include "kernel/parallel/threading/task_uitilities.h"
+#include "kernel/parallel/parallel_algos/reduction_operations.h"
+#include "kernel/parallel/utilities/result_holder.h"
 
 #include <boost/noncopyable.hpp>
 #include <stdexcept>
