+++ kernel/include/kernel/parallel/parallel_algos/linear_algebra/matrix_vector_product.h
@@ -2,10 +2,10 @@
 #define MATRIX_VECTOR_PRODUCT_H
 
 #include "kernel/base/types.h"
-#include "kernel/executors/simple_task.h"
-#include "kernel/executors/task_uitilities.h"
-#include "kernel/utilities/result_holder.h"
-#include "kernel/data_structs/range_1d.h"
+#include "kernel/parallel/threading/simple_task.h"
+#include "kernel/parallel/threading/task_uitilities.h"
+#include "kernel/parallel/utilities/result_holder.h"
+#include "kernel/utilities/range_1d.h"
 
 #include <boost/noncopyable.hpp>
